topik_03_Kira/computer-vision/requirements.txt
# Computer Vision Dependencies
opencv-python==4.8.1.78
mediapipe==0.10.8
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
Pillow==10.1.0

# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database and Storage
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
tqdm==4.66.1